extract get_ascii_colors from info.rs
