+++ src/onefetch/language.rs
@@ -1,5 +1,5 @@
 use {
-    crate::onefetch::error::*,
+    crate::onefetch::{error::*, utils::num_to_color},
     colored::Color,
     regex::Regex,
     std::collections::HashMap,
@@ -357,5 +357,34 @@ impl Language {
         }
 
         languages
+    }
+
+    pub fn get_ascii_colors(
+        ascii_language: &Option<Language>,
+        dominant_language: &Language,
+        ascii_colors: &[String],
+        true_color: bool,
+    ) -> Vec<Color> {
+        let language = if let Some(ascii_language) = ascii_language {
+            ascii_language
+        } else {
+            &dominant_language
+        };
+
+        let colors = language.get_colors(true_color);
+
+        let colors: Vec<Color> = colors
+            .iter()
+            .enumerate()
+            .map(|(index, default_color)| {
+                if let Some(color_num) = ascii_colors.get(index) {
+                    if let Some(color) = num_to_color(color_num) {
+                        return color;
+                    }
+                }
+                *default_color
+            })
+            .collect();
+        colors
     }
 }
